+++ docs/screenshots/stripes_screenshot.py
@@ -1,0 +1,6 @@
+from support.maker import make_app
+
+app = make_app("lots_of_links")
+
+if __name__ == "__main__":
+    app.run()